Create README.md
This is a tiny easy game, that name is 「Gomoku」. The game rule is Players alternate turns placing a stone of their color on an empty intersection. The winner is the first player to form an unbroken chain of five stones horizontally, vertically, or diagonally. the coding is C# - winform.
learning from the youtuber 「小山的教學平台」.